requirements.txt
ccxt
pandas
plotly
streamlit
requests
python-dotenv
streamlit-lightweight-charts

# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
# Note: uvloop removed due to macOS ARM compatibility issues (SIGSEGV crashes)
# uvicorn will use standard asyncio instead

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
